Factor common get_next_pass, fix #13
